Fix -Wconversion and -Wstringop-overflow build warnings
- ssh_session.c: cast FIONBIO to (long) in ioctlsocket calls to
  suppress sign-conversion warning (FIONBIO is 0x8004667e, which
  overflows a signed long without the cast)
- window.c: replace strncat with snprintf via a local exe_dir buffer
  to fix -Wstringop-overflow (bound equalled source length)

+++ src/term/ssh_session.c
@@ -101,7 +101,7 @@ int ssh_connect(SshSession *s, const char *host, int port) {
         /* Switch to non-blocking so we can timeout the connect. */
 #ifdef _WIN32
         u_long nb = 1;
-        ioctlsocket(sock, FIONBIO, &nb);
+        ioctlsocket(sock, (long)FIONBIO, &nb);
 #else
         int fl = fcntl(sock, F_GETFL, 0);
         fcntl(sock, F_SETFL, fl | O_NONBLOCK);
@@ -141,7 +141,7 @@ int ssh_connect(SshSession *s, const char *host, int port) {
             /* Restore blocking mode. */
 #ifdef _WIN32
             u_long nb2 = 0;
-            ioctlsocket(sock, FIONBIO, &nb2);
+            ioctlsocket(sock, (long)FIONBIO, &nb2);
 #else
             fcntl(sock, F_SETFL, fl & ~O_NONBLOCK);
 #endif
